Refactor: 중앙화된 타입 시스템으로 통합 - combination 컴포넌트들의 로컬 타입을 src/types/components/combination.ts로 통합 - data 폴더에서 타입과 데이터 분리 - 모든 컴포넌트에서 중앙화된 타입 시스템 사용 - 코드 중복 제거 및 타입 관리 일원화
